ble_ws_service/requirements.txt
bleak==2.0.0
pysnooz==0.10.0
websockets==15.0.1
PyYAML==6.0.3

